file relocations and import updates
