+++ frontend/src/graphql/user/queries.ts
@@ -8,12 +8,3 @@ export const LOGIN_QUERY = gql`
         }
     }
 `;
-
-export const GET_MY_PRODUCTS_QUERY = gql`
-    {
-        getMyProducts {
-            id
-            title
-        }
-    }
-`;
